fix: move sorting back to the selector
